Migrate unix_sockname fixture to conftest module
